+++ itinerary_planner/src/components/place-card/Placecard.jsx
