+++ itinerary_planner/src/components/split-text/SplitText.jsx
