+++ itinerary_planner/src/components/spotlight-card/SpotlightCard.css
